add impact question

+++ main.js
@@ -1,15 +1,25 @@
 const $ = document.querySelector.bind(document);
 const $$ = document.querySelectorAll.bind(document);
 
+const header =$('header')
 const addQuestion = $(".add-question");
 const btnEraser = $(".fa-eraser");
 const showCreator = $(".show-creator");
 const emptyQuestion = $(".empty-question");
 
-const btnddNewQuestion =$(".btnddNewQuestion")
-
+const btnAddNewQuestion =$(".btnAddNewQuestion")
+const page = $('.page');
 let checkQuestionLocal = localStorage.getItem('hiden');
 
+// --------------StickyNavbar----------------
+function stickyNavbar() {
+    header.classList.toggle("scrolled", window.pageYOffset > 0);
+
+};
+
+stickyNavbar();
+
+window.addEventListener("scroll", stickyNavbar);
 // ---------------HIDEN QUESTION ADD----------------
 
 //check is question
@@ -32,13 +42,12 @@ addQuestion.addEventListener("click", ()=> {
 });
 
 btnEraser.addEventListener("click", ()=> {
-    emptyQuestion.classList.remove("hiden")
-    showCreator.classList.add("hiden")
-    localStorage.setItem("hiden",0)
+    if (confirm('You are going to delete all elements in this survey. Do you want to proceed??')) {
+        emptyQuestion.classList.remove("hiden")
+        showCreator.classList.add("hiden")
+        localStorage.setItem("hiden",0)
+    }
+
 });
 
 // --------------- ADD NEW QUESION----------------
-
-btnAddNewQuestion.addEventListener("click", ()=> {
-
-})
